prompt undefined fix
Hello when i tried this grunt task i have this error blocking me :
Fatal error: Cannot read property 'prompt' of undefined
I had to patch out the script in order to have execution. 
After the patch i have outputs like this :
1=> "& did not start a character reference. (& probably should have been escaped as &amp;.)" Line no: undefined
(the typo was on the title tag in the html)
I don't know why this happened or if this patch could be useful to you but anyway i report it to you.
Thanks for your code.